+++ CHANGELOG.md
@@ -165,4 +165,14 @@ can get data for both organization and organization brand pages.
 * Extend `Post` object to include image details mapping
 * Add helper methods to `Post` to extract title, description and image URLs
 
-## 4.2.1 (Work in progress)
+## 4.2.1 (Feb 27, 2023)
+* Add missing LinkedIn annotations
+
+## 4.3.0 (Mar 8, 2023)
+* Add support to upload video files from URLs
+
+## 4.3.1 (Mar 14, 2023)
+* Add a flag to note that we want to upload a thumbnail image along with the video
+
+## 4.4.0 (Mar 16, 2023)
+* Add support to upload thumbnail image to a video